python/instrumentation/openinference-instrumentation-elevenlabs/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openinference-instrumentation-elevenlabs"
dynamic = ["version"]
description = "OpenInference ElevenLabs Instrumentation"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9, <3.15"
authors = [
  { name = "OpenInference Authors", email = "oss@arize.com" },
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
]
dependencies = [
  "opentelemetry-api",
  "opentelemetry-instrumentation",
  "opentelemetry-semantic-conventions",
  "openinference-instrumentation>=0.1.17",
  "openinference-semantic-conventions>=0.1.12",
  "wrapt",
]

[project.optional-dependencies]
instruments = [
  "elevenlabs>=1.0.0",
]
test = [
  "mypy>=1.18.2",
  "pytest-recording",
  "pytest-asyncio>=0.21.0",
  "elevenlabs",
  "opentelemetry-sdk>=1.20.0",
  "opentelemetry-exporter-otlp-proto-http",
]

[project.entry-points.opentelemetry_instrumentor]
elevenlabs = "openinference.instrumentation.elevenlabs:ElevenLabsInstrumentor"

[project.entry-points.openinference_instrumentor]
elevenlabs = "openinference.instrumentation.elevenlabs:ElevenLabsInstrumentor"

[project.urls]
Homepage = "https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-elevenlabs"

[tool.hatch.version]
path = "src/openinference/instrumentation/elevenlabs/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
]

[tool.hatch.build.targets.wheel]
packages = ["src/openinference"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = [
  "tests",
]

[tool.mypy]
strict = true
explicit_package_bases = true
exclude = [
  "examples",
  "dist",
  "sdist",
  "tests",
]

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "wrapt",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "E501"]
"tests/**/*.py" = ["E501"]  # Ignore line length in tests

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
force-single-line = false
